whats wrong with me

+++ javaScript/parse.js
@@ -1,5 +1,5 @@
 function getInnerTextById(id){
     const value = document.getElementById(id).innerText;
-    const convertedValue = parseInt(value);
+    const convertedValue = parseFloat(value);
     return convertedValue;
 }
